Split actionOutput, variable and variables in their own files
