+++ __tests__/actions/runScriptOverSSH.spec.ts
@@ -33,4 +33,5 @@ describe('runScriptOverSSH function', () => {
 
     expect(actual).toEqual(expected);
   });
+
 });
