+++ src/utils/index.ts
@@ -1,19 +1,23 @@
+import { actionOutput } from './actionOutput';
 import { buildSerialization } from './buildSerialization';
 import { buildShortcut } from './buildShortcut';
 import { buildShortcutTemplate } from './buildShortcutTemplate';
 import { encodeShortcut } from './encodeShortcut';
 import { flatten } from './flatten';
 import { getItemType } from './getItemType';
+import { variable } from './variable';
 import { withActionOutput } from './withActionOutput';
 import { withVariables } from './withVariables';
 
 export {
+  actionOutput,
   buildSerialization,
   buildShortcut,
   buildShortcutTemplate,
   encodeShortcut,
   flatten,
   getItemType,
+  variable,
   withActionOutput,
   withVariables,
 };
